Move subprotocols traits to respective main files
