+++ latticefold/src/nifs/decomposition/structs.rs
@@ -1,18 +1,9 @@
 #![allow(non_snake_case, clippy::upper_case_acronyms)]
 use ark_serialize::{CanonicalDeserialize, CanonicalSerialize};
 use ark_std::marker::PhantomData;
-use lattirust_ring::{OverField, Ring};
+use lattirust_ring::Ring;
 
-use crate::{
-    arith::{Witness, CCS, LCCCS},
-    ark_base::*,
-    commitment::AjtaiCommitmentScheme,
-    commitment::Commitment,
-    decomposition_parameters::DecompositionParams,
-    nifs::error::DecompositionError,
-    transcript::Transcript,
-};
-use cyclotomic_rings::rings::SuitableRing;
+use crate::{ark_base::*, commitment::Commitment};
 
 /// The proof structure of the decomposition subprotocol.
 #[derive(Clone, Debug, PartialEq, CanonicalSerialize, CanonicalDeserialize)]
@@ -43,32 +34,6 @@ pub struct DecompositionProof<const C: usize, NTT: Ring> {
     pub x_s: Vec<Vec<NTT>>,
     /// Commitments to the decomposed witnesses.
     pub y_s: Vec<Commitment<C, NTT>>,
-}
-
-pub trait DecompositionProver<NTT: SuitableRing, T: Transcript<NTT>> {
-    fn prove<const W: usize, const C: usize, P: DecompositionParams>(
-        cm_i: &LCCCS<C, NTT>,
-        wit: &Witness<NTT>,
-        transcript: &mut impl Transcript<NTT>,
-        ccs: &CCS<NTT>,
-        scheme: &AjtaiCommitmentScheme<C, W, NTT>,
-    ) -> Result<
-        (
-            Vec<LCCCS<C, NTT>>,
-            Vec<Witness<NTT>>,
-            DecompositionProof<C, NTT>,
-        ),
-        DecompositionError,
-    >;
-}
-
-pub trait DecompositionVerifier<NTT: OverField, T: Transcript<NTT>> {
-    fn verify<const C: usize, P: DecompositionParams>(
-        cm_i: &LCCCS<C, NTT>,
-        proof: &DecompositionProof<C, NTT>,
-        transcript: &mut impl Transcript<NTT>,
-        ccs: &CCS<NTT>,
-    ) -> Result<Vec<LCCCS<C, NTT>>, DecompositionError>;
 }
 
 pub struct LFDecompositionProver<NTT, T> {
